+++ src/main/java/com/ngtesting/platform/model/SysNums.java
@@ -5,13 +5,13 @@ import java.io.Serializable;
 public class SysNums implements Serializable {
     private static final long serialVersionUID = -9140817575372640340L;
 
-    private Long key;
+    private Integer key;
 
-    public Long getKey() {
+    public Integer getKey() {
         return key;
     }
 
-    public void setKey(Long key) {
+    public void setKey(Integer key) {
         this.key = key;
     }
 }
